+++ adapters/apacdex/apacdex_test.go
@@ -1,4 +1,4 @@
-package valueimpression
+package apacdex
 
 import (
 	"testing"
@@ -9,12 +9,12 @@ import (
 )
 
 func TestJsonSamples(t *testing.T) {
-	bidder, buildErr := Builder(openrtb_ext.BidderValueImpression, config.Adapter{
+	bidder, buildErr := Builder(openrtb_ext.BidderApacdex, config.Adapter{
 		Endpoint: "//host"})
 
 	if buildErr != nil {
 		t.Fatalf("Builder returned unexpected error %v", buildErr)
 	}
 
-	adapterstest.RunJSONBidderTest(t, "valueimpressiontest", bidder)
+	adapterstest.RunJSONBidderTest(t, "apacdextest", bidder)
 }
