+++ adapters/apacdex/apacdextest/exemplary/banner-and-video.json
@@ -92,7 +92,7 @@
           "id": "test-request-id",
           "seatbid": [
             {
-              "seat": "valueimpression",
+              "seat": "apacdex",
               "bid": [
                 {
                   "id": "8ee514f1-b2b8-4abb-89fd-084437d1e800",
