+++ adapters/apacdex/apacdextest/supplemental/explicit-dimensions.json
