+++ adapters/apacdex/apacdextest/supplemental/invalid-response-no-bids.json
