+++ adapters/apacdex/apacdextest/supplemental/server-error-code.json
@@ -46,7 +46,7 @@
 
     "expectedMakeBidsErrors": [
       {
-        "value": "Bad server response: HTTP status 500",
+        "value": "Unexpected status code: 500. Run with request.debug = 1 for more info.",
         "comparison": "literal"
       }
     ]
