+++ adapters/apacdex/apacdextest/supplemental/wrong-impression-ext.json
@@ -19,7 +19,7 @@
 
   "expectedMakeRequestsErrors": [
     {
-      "value": "json: cannot unmarshal number into Go struct field ExtImpValueImpression.siteId of type string",
+      "value": "json: cannot unmarshal number into Go struct field ExtImpApacdex.siteId of type string",
       "comparison": "literal"
     }
   ]
