+++ openrtb_ext/imp_apacdex.go
@@ -1,0 +1,7 @@
+package openrtb_ext
+
+type ExtImpApacdex struct {
+	PlacementID string  `json:"placementId"`
+	SiteID      string  `json:"siteId"`
+	FloorPrice  float64 `json:"floorPrice"`
+}
