--- openrtb_ext/imp_valueimpression.go
@@ -1,5 +1,0 @@
-package openrtb_ext
-
-type ExtImpValueImpression struct {
-	SiteId string `json:"siteId"`
-}
