+++ static/bidder-params/apacdex.json
@@ -1,0 +1,32 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "title": "Apacdex Adapter Params",
+  "description": "A schema which validates params accepted by the Apacdex adapter",
+  "type": "object",
+  "properties": {
+    "placementId": {
+      "type": "string",
+      "description": "Placement ID provided by Apacdex"
+    },
+    "siteId": {
+      "type": "string",
+      "description": "Publisher site ID from Apacdex"
+    },
+    "floorPrice": {
+      "type": "number",
+      "description": "CPM bidfloor in USD"
+    }
+  },
+  "oneOf": [
+    {
+      "required": [
+        "placementId"
+      ]
+    },
+    {
+      "required": [
+        "siteId"
+      ]
+    }
+  ]
+}
